+++ models/userBalanceRecord.go
@@ -10,9 +10,9 @@ const TableNameUserBalanceRecode = "user_balance_record"
 // UserBalanceRecord 用户余额变动记录表（账变记录表）
 type UserBalanceRecord struct {
 	RecordId        int64           `json:"record_id" gorm:"record_id,primaryKey"`      // 主键ID
-	MerchantId      int             `json:"merchant_id" gorm:"merchant_id"`             // 商户ID
+	MerchantId      int64           `json:"merchant_id" gorm:"merchant_id"`             // 商户ID
 	UserId          int64           `json:"user_id" gorm:"user_id"`                     // 用户ID
-	PlatformId      int             `json:"platform_id" gorm:"platform_id"`             // 平台ID
+	PlatformId      int64           `json:"platform_id" gorm:"platform_id"`             // 平台ID
 	TransactionType int             `json:"transaction_type" gorm:"transaction_type"`   // 交易类型：1=转入游戏,2=从游戏转出,3=投注,4=结算派奖,5=押金,6=返还押金,7=取消投注,8=结算撤单,9=重新派奖,10=调整金额
 	Amount          decimal.Decimal `json:"amount" gorm:"amount"`                       // 变动金额（正数表示增加，负数表示减少）
 	BalanceBefore   decimal.Decimal `json:"balance_before" gorm:"balance_before"`       // 变动前余额
